LayoutTests/fast/custom-elements/scoped-custom-element-registry-disabled-expected.txt

This tests that attachShadow does not call "registry" attribute getter when scoped custom element registry is disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS registryGetterWasCalled is false
PASS successfullyParsed is true

TEST COMPLETE
